extra security for editing reesources

+++ controllers/bootcamps.js
@@ -101,16 +101,28 @@ exports.getBootcampsInRadius = ({ bootcampRepo }) =>
  */
 exports.updateBootcamp = ({ bootcampRepo }) =>
   asyncHandler(async (req, res, next) => {
-    const updatedBootcamp = await bootcampRepo.updateBootcamp(
-      req.params.id,
-      req.body
-    );
+    const bootcamp = await bootcampRepo.getBootcamp(req.params.id);
 
-    if (!updatedBootcamp) {
+    if (!bootcamp) {
       return next(
         new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
       );
     }
+
+    // Make sure user is bootcamp owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== "Admin") {
+      return next(
+        new ErrorResponse(
+          `User ${req.user.id} is not authorized to update this resource`,
+          401
+        )
+      );
+    }
+
+    const updatedBootcamp = await bootcampRepo.updateBootcamp(
+      req.params.id,
+      req.body
+    );
 
     res.status(200).json({ succes: true, data: updatedBootcamp });
   });
@@ -123,13 +135,25 @@ exports.updateBootcamp = ({ bootcampRepo }) =>
  */
 exports.deleteBootcamp = ({ bootcampRepo }) =>
   asyncHandler(async (req, res, next) => {
-    const deletedBootcamp = await bootcampRepo.deleteBootcamp(req.params.id);
+    const bootcamp = await bootcampRepo.getBootcamp(req.params.id);
 
-    if (!deletedBootcamp) {
+    if (!bootcamp) {
       return next(
         new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
       );
     }
+
+    // Make sure user is bootcamp owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== "Admin") {
+      return next(
+        new ErrorResponse(
+          `User ${req.user.id} is not authorized to delete this resource`,
+          401
+        )
+      );
+    }
+
+    await bootcampRepo.deleteBootcamp(req.params.id);
 
     res.status(200).json({ succes: true, data: {} });
   });
@@ -147,6 +171,16 @@ exports.uploadBootcampPhoto = ({ bootcampRepo }) =>
     if (!bootcamp) {
       return next(
         new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
+      );
+    }
+
+    // Make sure user is bootcamp owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== "Admin") {
+      return next(
+        new ErrorResponse(
+          `User ${req.user.id} is not authorized to update this resource`,
+          401
+        )
       );
     }
